keep local context menu after saving content library
